Changed send_metric_tm task to tasks
