+++ python_metrics_client/tasks.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from celery.utils.log import get_task_logger
 from celery import shared_task
-
+from pyramid_transactional_celery import task_tm
 from python_metrics_client.influx import send_data as actual_send_data
 from python_metrics_client.metrics_client import send_metric as actual_send_metric, send_product_metric
 
@@ -36,7 +36,6 @@ def send_metric(self, environment, metric, value, tags, timestamp=None, server=N
     username = self.app.conf.metrics_user
     password = self.app.conf.metrics_user
 
-
     actual_send_metric(server, port, environment, metric, value, tags, timestamp, client_type, username, password)
 
 
@@ -53,5 +52,20 @@ def send_product_metric(self, environment, product, metric, value, tags, timesta
     username = self.app.conf.metrics_user
     password = self.app.conf.metrics_user
 
+    send_product_metric(server, port, environment, product,  metric, value, tags, timestamp, client_type,
+                        username, password)
 
-    send_product_metric(server, port, environment, product,  metric, value, tags, timestamp, client_type, username, password)
+
+@task_tm(bind=True, queue='metrics_client')
+def send_metric_tm(self, environment, metric, value, tags, timestamp=None, server=None, port=None):
+
+    logger.info('Sending metric {}'.format(metric))
+    if not server:
+        server = self.app.conf.metrics_server
+
+    if not port:
+        port = self.app.conf.metrics_client_port
+
+    client_type = self.app.conf.metrics_client_type
+
+    actual_send_metric(server, port, environment, metric, value, tags, timestamp, client_type)
